FIX: Pixelmap was not drawn correctly
The issue was that only some pixels were drawn on the scene.
This fix solves this issue by commenting the line where OldPixelMap was assigned with the current content of PixelMap
and initializing OldPixelMap array with 0s. With this new initialization now the software knows there are new pixels
to be drawn.